Run the Plumber API

Run the Plumber API

Usage

run_api(.host = "0.0.0.0", .port = 8080, ...)

Arguments

.host

Host - defaults to "0.0.0.0" (localhost)

.port

Post - defaults to 8080

...

Arguments passed on to plumber::pr_run

pr

A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.

host

A string that is a valid IPv4 or IPv6 address that is owned by this server, which the application will listen on. "0.0.0.0" represents all IPv4 addresses and "::/0" represents all IPv6 addresses.

port

A number or integer that indicates the server port that should be listened on. Note that on most Unix-like systems including Linux and Mac OS X, port numbers smaller than 1025 require root privileges.

debug

If TRUE, it will provide more insight into your API errors. Using this value will only last for the duration of the run. If pr_set_debug() has not been called, debug will default to interactive() at pr_run() time

docs

Visual documentation value to use while running the API. This value will only be used while running the router. If missing, defaults to information previously set with pr_set_docs(). For more customization, see pr_set_docs() for examples.

swaggerCallback

An optional single-argument function that is called back with the URL to an OpenAPI user interface when one becomes ready. If missing, defaults to information set with pr_set_docs_callback(). This value will only be used while running the router.

quiet

If TRUE, don't print routine startup messages.

Value

Plumber API